// client/src/api/types.ts
export type CorrelationPoint = {
    factor: string
    r: number
}

export type AttendancePoint = {
    attendance: number
    avg_score: number
}

export type HoursStudiedPoint = {
    hours_studied: number
    avg_score: number
}

export type AccessToResourcesPoint = {
    access_to_resources: number
    avg_score: number
}

export type ParentalInvolvementPoint = {
    parental_involvement: number
    avg_score: number
}

export type TutoringSessionsPoint = {
    tutoring_sessions: number
    avg_score: number
}